Clean up Linux compiler warnings Add rhel4 binary directories to list to be ignored by cvs

+++ CalibData/src/Acd/AcdCalibObj.cxx
@@ -68,7 +68,7 @@ namespace CalibData {
   
   AcdCalibObj::AcdCalibObj(STATUS status, const std::vector<float>& vals, const AcdCalibDescription& desc) 
     :m_status(status){
-    assert(vals.size() <= desc.nVar());
+    assert((int) vals.size() <= desc.nVar());
     sizeVals(desc.nVar());   
     setVals(vals,status);
   }
